Add type assertion example to using extended state db within precompiled (#15)
contracts which can be used for accessing methods not found in the
standard core interface defined in this (precompile package) or core.

The interfaces could be extending to support a generic type with the
constraint that it implements StateDB, however; this forces us to
instantiate the type in all types that use the accessible state.

(cherry picked from commit 3ffaf102efbb263d182bf8425ddf3a007bfdf016)