chore: Remove stateful precompile implementation
This removes the core changes and precompile code added initially to
support stateful precompiles.

After much learning, prototyping, and writing E2E test cases for
stateful precompile logic, we are re-implementing a solution with a new
perspective -- correcting design mistakes and separating the stateful
precompile logic from the interface extending the EVM.

This re-implementation will provide a more secure and flexible
implementation with better separation of concerns and provide an
extension that is not coupled to the existing native precompile
implementation.